requirements.txt
boto3
ipdb
pyyaml
